We can play sounds when input are pressed
There is two scripts. The "TESTSoundManager" and "SoundManager" does the same thing, but the "SoundManager" has not only a list of sounds but a list of types of sounds, and inside of those types the songs. And the list resize automaticly in the inspector because of some crazy advanced stuff in the script (more complex than "TESTSoundManager")